Separated Atom class into separate file
